pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "dudubarbaro", email = "dudubarbaro"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]
